POSSIBLE REMOVE! Make sure apcupsd's scripts can find a "mail" program
This is a left-over I am leaving in for now. We can change APCUPSD_MAIL during
compilation, which is now being done. No need to provide a mail program that
provides the "mail" command, you can just provide one yourself to override the package.

The program to run MUST have the name mail. sendmail is not an option. Perhaps
we should patch it with a snippet in the package to allow other MTAs to be used?