only accept gexps as power event handlers
While program-file may be another option to accept, the problem is that
we need to have a way to call back to apccontrol, the script that handles the
incoming events and dispatches to the user-provided script first. The easiest
way to get that into what a user can provide is through splicing the user's
provided gexp into another which sets the apccontrol script's path.